wip: Addresses @hristoterezov's comments.
