+++ modules/util/helpers.js
@@ -24,24 +24,3 @@ export function createDeferred() {
 export function createTaskQueue() {
     return new TaskQueue();
 }
-
-/**
- * Downloads a JSON object.
- *
- * @param json the JSON object to download.
- * @param filename the filename to give to the downloaded file.
- */
-export function downloadJSON(json, filename) {
-    const data = encodeURIComponent(JSON.stringify(json, null, '  '));
-
-    const elem = document.createElement('a');
-
-    elem.download = filename;
-    elem.href = `data:application/json;charset=utf-8,\n${data}`;
-    elem.dataset.downloadurl = [ 'text/json', elem.download, elem.href ].join(':');
-    elem.dispatchEvent(new MouseEvent('click', {
-        view: window,
-        bubbles: true,
-        cancelable: false
-    }));
-}
