--- react/features/connection-indicator/actionTypes.js
@@ -1,8 +1,0 @@
-/**
- * Create an action for saving the conference logs.
- *
- * {
- *     type: CONNECTION_INDICATOR_SAVE_LOGS
- * }
- */
-export const CONNECTION_INDICATOR_SAVE_LOGS = 'CONNECTION_INDICATOR_SAVE_LOGS';
