+++ react/features/connection-indicator/middleware.js
